+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @interop/data-integrity-core Changelog
 
-## 6.0.0 - 
+## 6.0.0 - 2026-05-31
 ### Changed
 - **BREAKING**: Fork from `@digitalcredentials/ssi@5.5.0` to `@interop/data-integrity-core@6.0.0`.